% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphiques.R
\name{assembler}
\alias{assembler}
\title{Assemble les series demandees en un tableau long pret a tracer.}
\usage{
assembler(con, series, debut = NULL, fin = NULL)
}
\arguments{
\item{series}{liste de listes (code_interne, frequence, pays)}
}
\description{
Assemble les series demandees en un tableau long pret a tracer.
}
