% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collecte.R
\name{charger_pays}
\alias{charger_pays}
\title{La liste des pays vient de l'API de la Banque mondiale, qui fournit region,
groupe de revenu et code ISO en un seul appel.}
\usage{
charger_pays(con)
}
\description{
La liste des pays vient de l'API de la Banque mondiale, qui fournit region,
groupe de revenu et code ISO en un seul appel.
}
