% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collecte.R
\name{collecter_indicateur}
\alias{collecter_indicateur}
\title{Collecte un indicateur et l'ecrit en base.}
\usage{
collecter_indicateur(con, ligne, debut = NULL, fin = NULL)
}
\value{
liste (creees, modifiees)
}
\description{
Collecte un indicateur et l'ecrit en base.
}
