% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collecte.R
\name{initialiser_base}
\alias{initialiser_base}
\title{Charge le catalogue et la liste des pays}
\usage{
initialiser_base(con, avec_pays = TRUE)
}
\arguments{
\item{con}{connexion ouverte.}

\item{avec_pays}{recharger aussi la liste des pays depuis la Banque mondiale.}
}
\description{
Charge le catalogue et la liste des pays
}
