% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteurs.R
\name{periode_vers_date}
\alias{periode_vers_date}
\title{Normalise une periode au premier jour de la periode couverte.}
\usage{
periode_vers_date(p)
}
\description{
Cette convention est ce qui permet de superposer sur un meme graphique des
series de pas differents. Renvoie une liste (date, frequence).
}
