% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Lance la plateforme OPESc+}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{fonction executee avant le premier appel de l'application.}

\item{options}{options passees a [shiny::shinyApp()].}

\item{enableBookmarking}{mode de mise en signet.}

\item{uiPattern}{expression reguliere des chemins servis.}

\item{...}{options accessibles ensuite par `golem::get_golem_options()`.}
}
\description{
Lance la plateforme OPESc+
}
\examples{
\dontrun{
opescplus::run_app()
}
}
