% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphiques.R
\name{tracer}
\alias{tracer}
\title{Trace les series. `type` vaut "ligne", "barre" ou "aire".}
\usage{
tracer(d, type = "ligne", base100 = FALSE, titre = NULL)
}
\description{
Trace les series. `type` vaut "ligne", "barre" ou "aire".
}
